requirements.txt
alembic==0.9.5
appnope==0.1.0
click==6.7
decorator==4.1.2
Flask==0.12.2
Flask-Admin==1.5.0
-e git+git@github.com:frostming/Flask-BlogTheme.git#egg=Flask_BlogTheme
Flask-Migrate==2.1.1
Flask-Moment==0.5.1
Flask-SQLAlchemy==2.2
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
mistune==0.7.4
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
PyYAML==3.12
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.13
traitlets==4.3.2
Unidecode==0.4.21
wcwidth==0.1.7
Werkzeug==0.12.2
WTForms==2.1
